Fix old Go non-existent build path error.
If you specify just an output path without a filename when building the app, it will say that "build output dir/ already exists and is a directory" then it'll crash the build process.
This is considered as a bug (https://github.com/golang/go/issues/41313) and was fixed in Go 1.16. The fix is for older versions.